LPS-93819 Move ClassTypeReaders from *.asset to *.asset.model packages
