+++ modules/apps/dynamic-data-lists/dynamic-data-lists-web/src/main/java/com/liferay/dynamic/data/lists/web/internal/asset/model/DDLRecordSetClassTypeReader.java
@@ -12,13 +12,14 @@
  * details.
  */
 
-package com.liferay.dynamic.data.lists.web.internal.asset;
+package com.liferay.dynamic.data.lists.web.internal.asset.model;
 
 import com.liferay.asset.kernel.model.ClassType;
 import com.liferay.asset.kernel.model.ClassTypeReader;
 import com.liferay.dynamic.data.lists.model.DDLRecordSet;
 import com.liferay.dynamic.data.lists.model.DDLRecordSetConstants;
 import com.liferay.dynamic.data.lists.service.DDLRecordSetServiceUtil;
+import com.liferay.dynamic.data.lists.web.internal.asset.DDLRecordSetClassType;
 import com.liferay.portal.kernel.exception.PortalException;
 import com.liferay.portal.kernel.util.LocaleUtil;
