+++ portal-impl/src/com/liferay/portlet/documentlibrary/asset/model/DLFileEntryClassTypeReader.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.portlet.documentlibrary.asset;
+package com.liferay.portlet.documentlibrary.asset.model;
 
 import com.liferay.asset.kernel.model.ClassType;
 import com.liferay.asset.kernel.model.ClassTypeReader;
@@ -23,6 +23,7 @@ import com.liferay.document.library.kernel.service.DLFileEntryTypeServiceUtil;
 import com.liferay.portal.kernel.exception.PortalException;
 import com.liferay.portal.kernel.language.LanguageUtil;
 import com.liferay.portal.kernel.util.LocaleUtil;
+import com.liferay.portlet.documentlibrary.asset.DLFileEntryClassType;
 
 import java.util.ArrayList;
 import java.util.List;
